Crate token

A simple string tokenizer, and a tokenizer of sentences based on it. The tokenizer ignores all the given separator chars, and returns the characters between as string slice.

Examples

General use (as an iterator)

This is how you will probably use it

let separators = vec![' ', '\n', '\t', '\r'];
let source: &str = "    Hello world \n  How do you do\t-Finely I hope";

let mut tokenizer = token::Tokenizer::new(source.as_bytes(), separators);
println!("Tokenizing...");
for token in tokenizer {
    println!("- Got token: {}", token.unwrap());
}
println!("Done!");

Behavior

This is what to expect when parsing a string (or input from a reader)

let separators = vec![' ', '\n', '\t', '\r'];
let source: &str = "    Hello world \n  How do you do\t-Finely I hope";

let mut tokenizer = token::Tokenizer::new(source.as_bytes(), separators);
assert_eq!("Hello",     tokenizer.next().expect("1").unwrap());
assert_eq!("world",     tokenizer.next().expect("2").unwrap());
assert_eq!("How",       tokenizer.next().expect("3").unwrap());
assert_eq!("do",        tokenizer.next().expect("4").unwrap());
assert_eq!("you",       tokenizer.next().expect("5").unwrap());
assert_eq!("do",        tokenizer.next().expect("6").unwrap());
assert_eq!("-Finely",   tokenizer.next().expect("7").unwrap());
assert_eq!("I",         tokenizer.next().expect("8").unwrap());
assert_eq!("hope",      tokenizer.next().expect("9").unwrap());
assert_eq!(None,        tokenizer.next());

Structs

SentenceSplitter
A structure for iteratively splitting stringy things into sentences

Tokenizer
A tokenizer returning string slices from a reader