Struct token::Tokenizer

pub struct Tokenizer<R: Read> {
    // some fields omitted
}

A tokenizer returning string slices from a reader

Note: The returned tokens are only valid until the next iteration

Methods

impl<R> Tokenizer<R> where R: Read

fn new(reader: R, separators: Vec<char>) -> Tokenizer<R>

Creates a new tokenizer from a reader and a set of separating characters

let terminators = vec![' ', '\n', '\t'];
let source: &str = "   Hello world\nHow do you do\t-Finely I hope";

let mut tokenizer = token::Tokenizer::new(source.as_bytes(), terminators);